various fixups
1) cleanup buildrpm for 2.3.6 and for x86
2) get PLATFORM_VERSION and TARGET_ARCH from build scripts via printvar.mk
3) work around weird bash bug where script fails after a 'false' if statement (the higher recursion level thinks that the entire script has failed!)
4) Separated off the 'additions to manifest.xml' as 'new_projects.xml', which is appended to the manifest by strip-projects.sh
(after it strips off the 'unneeded' projects, it adds back in these new ones)

+++ scripts/fixup_common.sh
@@ -1,4 +1,5 @@
-#
+#!/bin/bash
+set -x
 mkdir libnativehelper
 ln -s ../dalvik/libnativehelper/include libnativehelper/
 gzip frameworks/base/core/jni/Android.mk
@@ -55,35 +56,33 @@ if [ -e vendor_extra/qcom/proprietary ] ; then
     ln -s ../../vendor_extra/qcom/proprietary vendor/qcom/proprietary
     sed -i.001 -e "/BOARD_USE_QCOM_LLVM_CLANG_RS/d" device/qcom/msm8960/BoardConfig.mk
     sed -i.001 -e "/llvm-select.mk/d" device/qcom/common/common.mk
-    sed -i.001 -e "/\.apk/s/^/#/" vendor_extra/qcom/proprietary/common/config/device-vendor.mk
     sed -i.001 \
         -e "/^LOCAL_MODULE_SUFFIX.*.apk/,/BUILD_PREBUILT/s/^/#/" \
         -e "/^LOCAL_MODULE_SUFFIX.*(COMMON_JAVA_PACKAGE_SUFFIX)/,/BUILD_PREBUILT/s/^/#/" \
         vendor_extra/qcom/proprietary/prebuilt_HY11/target/product/msm8960/Android.mk
+    if [ -e vendor_extra/qcom/proprietary/clang-rs ] ; then
+        gzip vendor_extra/qcom/proprietary/clang-rs/clang-host-build.mk
+        find vendor_extra/qcom/proprietary/clang-rs/ -name Android.mk -exec gzip {} \;
+    fi
+
     if [ -e vendor_extra/qcom/proprietary/flash10-bin ] ; then
         ls vendor_extra/qcom/proprietary/flash10-bin/*/Android.mk | while read filename ; do
             gzip $filename
         done
     fi
     [ -f vendor_extra/qcom/proprietary/neocore/Android.mk ] && gzip vendor_extra/qcom/proprietary/neocore/Android.mk
-    gzip vendor_extra/qcom/proprietary/wfd/wdsm/service/jni/Android.mk
+    [ -f device/qcom/common/generate_extra_images.mk ] && \
+        sed -i.001 -e "/^RECOVERY_FROM_BOOT_PATCH/d" device/qcom/common/generate_extra_images.mk
+    [ -f vendor_extra/qcom/proprietary/mm-http/Android.mk ] && \
+        sed -i.001 -e "/include/d" vendor_extra/qcom/proprietary/mm-http/Android.mk
+    [ -e prebuilts/ndk ] && (cd prebuilt/ndk; ln -s ../../prebuilts/ndk/* .)
+    # 2.3.6
+    [ -f vendor_extra/qcom/proprietary/common/build/vendorsetup.sh ] && \
+        chmod a+x vendor_extra/qcom/proprietary/common/build/vendorsetup.sh
     sed -i.001 -e "s/ ndk\// prebuilt\/ndk\/android-ndk-r6\//" vendor_extra/qcom/proprietary/wfd/rtsp/Android.mk
     sed -i.001 -e "/FEATURE_GSIFF_ANDROID_NDK = 1/d" vendor_extra/qcom/proprietary/gps/isagnav/gsiff/Android.mk
-    if [ -e vendor_extra/qcom/proprietary/clang-rs ] ; then
-        gzip vendor_extra/qcom/proprietary/clang-rs/clang-host-build.mk
-        find vendor_extra/qcom/proprietary/clang-rs/ -name Android.mk -exec gzip {} \;
-    fi
-    [ -f vendor/qcom/opensource/bt-wlan-coex/btc/wlan_btc_usr_svc.c ] && \
-        sed -i.001 -e "s/ LOGI/ ALOGI/" -e "s/ LOGE/ ALOGE/" -e "s/ LOG_FATAL/ ALOG_FATAL/" \
-        vendor/qcom/opensource/bt-wlan-coex/btc/wlan_btc_usr_svc.c
-    [ -f vendor/qcom/opensource/bt-wlan-coex/btces/btces_plat.h ] && \
-        sed -i.001 -e "s/ LOGI/ ALOGI/" -e "s/ LOGE/ ALOGE/" -e "s/ LOG_FATAL/ ALOG_FATAL/" \
-        vendor/qcom/opensource/bt-wlan-coex/btces/btces_plat.h
-    sed -i.001 -e "/^RECOVERY_FROM_BOOT_PATCH/d" device/qcom/common/generate_extra_images.mk
-    sed -i.001 -e "/include/d" vendor_extra/qcom/proprietary/mm-http/Android.mk
-    (cd prebuilt/ndk; ln -s ../../prebuilts/ndk/* .)
-    # 2.3.6
-    #[ -f vendor/qcom/android-open/libopencorehw/Android.mk ] && gzip vendor/qcom/android-open/libopencorehw/Android.mk
-    chmod a+x vendor/qcom/proprietary/common/build/vendorsetup.sh
-#sed -i.001 -e "s/external\/opencore\/extern_libs_v2\/khronos\/openmax\/include/vendor\/qcom\/opensource\/omx\/mm-core\/omxcore\/inc/" vendor/qcom/proprietary/mm-video/DivxDrmDecrypt/Android.mk
+    gzip vendor_extra/qcom/proprietary/wfd/wdsm/service/jni/Android.mk
+    sed -i.001 -e "/\.apk/s/^/#/" vendor_extra/qcom/proprietary/common/config/device-vendor.mk
 fi
+#bash bug: don't end the file with a conditional
+true
